APPS-3024: Fixed comemnts after AA review
