+++ tests/_data/recipes/src/integrator-recipes-master/TestIntegratorWirePlugin/1.0.2/installer-manifest.json
@@ -2,8 +2,22 @@
     "wire-plugin": [
         {
             "target": "\\Pyz\\Zed\\TestIntegratorWirePlugin\\TestIntegratorWirePluginDependencyProvider::getTestPlugins",
-            "source": "\\Spryker\\Zed\\TestIntegratorWirePlugin\\Communication\\Plugin\\TestIntegratorWirePlugin",
+            "source": "\\Spryker\\Zed\\TestIntegratorWirePlugin\\Communication\\Plugin\\TestIntegratorWirePlugin"
+        },
+        {
+            "target": "\\Pyz\\Zed\\TestIntegratorWirePlugin\\TestIntegratorWirePluginDependencyProvider::getTestPlugins",
+            "source": "\\Spryker\\Zed\\TestIntegratorWirePlugin\\Communication\\Plugin\\TestIntegratorWirePluginIndex",
             "index": "\\Spryker\\Zed\\TestIntegratorWirePlugin\\TestIntegratorWirePluginConfig::TEST_INTEGRATOR_WIRE_PLUGIN"
+        },
+        {
+            "target": "\\Pyz\\Zed\\TestIntegratorWirePlugin\\TestIntegratorWirePluginDependencyProvider::getTestPlugins",
+            "source": "\\Spryker\\Zed\\TestIntegratorWirePlugin\\Communication\\Plugin\\TestIntegratorWirePluginStaticIndex",
+            "index": "static::TEST_INTEGRATOR_WIRE_PLUGIN_STATIC_INDEX"
+        },
+        {
+            "target": "\\Pyz\\Zed\\TestIntegratorWirePlugin\\TestIntegratorWirePluginDependencyProvider::getTestPlugins",
+            "source": "\\Spryker\\Zed\\TestIntegratorWirePlugin\\Communication\\Plugin\\TestIntegratorWirePluginStringIndex",
+            "index": "TEST_INTEGRATOR_WIRE_PLUGIN_STRING_INDEX"
         }
     ]
 }
